+++ packages/devtools-evm/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @layerzerolabs/devtools-evm
 
+## 0.3.10
+
+### Patch Changes
+
+- 8b32a79: Use less strict version requirements for layerzero peer dependencies
+- Updated dependencies [8b32a79]
+  - @layerzerolabs/devtools@0.3.15
+  - @layerzerolabs/io-devtools@0.1.10
+
+## 0.3.9
+
+### Patch Changes
+
+- 8935369: Add experimental support for batched sending
+- c13cee5: Use getNextNonce when proposing transactions to gnosis
+- Updated dependencies [8935369]
+  - @layerzerolabs/devtools@0.3.13
+
 ## 0.3.8
 
 ### Patch Changes